Use folder name as flow name in the trace
